DocLens/client/requirements.txt
# Frontend dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
